数据库迁移：Neon Postgres → Cloudflare D1（边缘 SQLite）
- 新增 functions/_d1.js：D1 适配层（neon 风格 sql 模板 → D1 prepare/bind/all），data 列 JSON 自动 parse
- functions/_utils.js 重写为 D1，schema 改为 SQLite DDL
- 全部 functions 方言改造：去 ::cast、now()→strftime、ANY/FILTER/ILIKE、错误码、去 CASCADE
- 新增 wrangler.toml（D1 binding=DB、R2 binding=R2）与本地迁移脚本

+++ functions/api/polls/[id]/index.js
@@ -3,7 +3,7 @@
 import { getSql, corsResponse, optionsResponse } from '../../../_utils.js'
 
 async function getVoteCounts(sql, pollId) {
-  const result = await sql`SELECT option_id, COUNT(*)::int as votes FROM poll_votes WHERE poll_id = ${pollId} GROUP BY option_id`
+  const result = await sql`SELECT option_id, COUNT(*) as votes FROM poll_votes WHERE poll_id = ${pollId} GROUP BY option_id`
   const counts = {}
   for (const row of result) counts[row.option_id] = row.votes
   return counts
